requirements.txt
Django==4.2
djangorestframework==3.12